feat: added domain entity models, character creation and search ref: changed user and character db model, updated controllers for better type safety, minor updates and fixes

+++ backend/src/services/auth.service.ts
@@ -8,7 +8,7 @@ import UserMapper from "../mappers/user.mapper.js"
 import { comparePassword, hashPassword } from "../utils/hash.js"
 import JWTHelper from "../utils/jwt.helper.js"
 import { ConflictError, UnauthorizedError } from "../errors/app.error.js"
-import type { RefreshTokenCreateDTO } from "../types/token.js"
+import type { CreateUserInput } from "../types/user.js"
 
 class AuthService {
   constructor(
@@ -21,17 +21,18 @@ class AuthService {
       throw new ConflictError("User with this email already exists")
     }
     const hashedPassword = await hashPassword(userData.password)
-    const user = await this.userRepository.create({
-      ...userData,
-      password: hashedPassword,
-    })
+    const createInput: CreateUserInput = {
+      email: userData.email,
+      name: userData.name,
+      passwordHash: hashedPassword,
+    }
+    const user = await this.userRepository.create(createInput)
     const dbToken = await this.refreshTokenRepository.create({
       sub: user.id,
     })
     const { access, refresh } = JWTHelper.generateTokens({
       sub: user.id,
       jti: dbToken.jti,
-      role: user.role,
     })
 
     return {
@@ -43,10 +44,7 @@ class AuthService {
   async login(userData: UserLoginDTO): Promise<AuthResponseDTO> {
     const user = await this.userRepository.getByEmail(userData.email)
     if (!user) throw new UnauthorizedError("Invalid credentials")
-    const isPasswordCorrect = await comparePassword(
-      userData.password,
-      user.password,
-    )
+    const isPasswordCorrect = await comparePassword(userData.password, user.passwordHash)
     if (!isPasswordCorrect) throw new UnauthorizedError("Invalid credentials")
     const dbToken = await this.refreshTokenRepository.create({
       sub: user.id,
@@ -54,7 +52,6 @@ class AuthService {
     const { access, refresh } = JWTHelper.generateTokens({
       sub: user.id,
       jti: dbToken.jti,
-      role: user.role,
     })
 
     return { user: UserMapper.toPublicDto(user), access, refresh }
@@ -64,8 +61,7 @@ class AuthService {
     if (!decoded || !decoded.jti || !decoded.sub) throw new UnauthorizedError()
     const oldDbToken = await this.refreshTokenRepository.getByJti(decoded.jti)
     if (!oldDbToken) {
-      if (decoded.sub)
-        await this.refreshTokenRepository.deleteAllBySub(parseInt(decoded.sub))
+      if (decoded.sub) await this.refreshTokenRepository.deleteAllBySub(parseInt(decoded.sub))
       throw new UnauthorizedError()
     }
     const user = await this.userRepository.getById(oldDbToken.sub)
@@ -79,7 +75,6 @@ class AuthService {
 
     const { access, refresh } = JWTHelper.generateTokens({
       jti: newDbToken.jti,
-      role: user.role,
       sub: user.id,
     })
